add drag bar
There's a slight hiccup that I couldn't solve and so I implemented a workaround. The way it works is that the scroll bar follows the mouse and so when the user released it stops dragging. The issue is that because the page is more than just the two inputs and the scrollbar, the height it calculates is slightly off. This would cause the user to be unable to stop dragging. I attempted to try and correctly calculate the height but to no avail. Instead I wrote a workaround so that when it jumps the user can still drag and not be stuck dragging. When you fit the page so that only the input fields are in the viewport, it will work perfectly, meaning the scrollbar and the mouse are attached to each other. Otherwise, the mouse and the scrollbar will be slightly off from each other.